+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed:
 - more consistent bottom sheet modals
 - remove limit in tag search results
+- replace monospace font with main font & tabularFigures font feature
 
 ## [0.8.275] - 2023-02-26
 ### Changed: